Converted int to uint8 where applicable, and made brute force compare more cache friendly.
Still fails to pass for 500000+ vectors. While comparison seem to be
very fast, the correlated pair of similarity 70+ seem to wind up in
different subbuckets more often than not for large instances. This
should not be the case for relatively low amounts of subbuckets
indicating a bug might be present.